completed the required changes

+++ Main.java
@@ -2,7 +2,7 @@ public class Main {
     public static void main(String[] args) {
         Restaurant.displayBurgersMenu();
 
-        Topping t = new Topping("Tomato",0.27);
-        BasicBurger b= new BasicBurger("BasicBurger with White with Chicken",3.56,4,"Chicken");
+        Topping t = new Topping("Tomato",(Double)0.27);
+        BasicBurger b= new BasicBurger("BasicBurger with White with Chicken",(Double)3.56,4,"Chicken");
     }
 }
